Actually make the three panels relate to each other
This is surprisingly difficult.

Midjourney cannot do this.

DALL-E v3 *definitely* cannot do this natively, with or without the help of GPT Vision.

So what we do is we be a bit tricksy, and we carefully prompt DALL-E to create
a *single* picture comprised of four sub-pictures. But in order to do this, we
must create a phantom fourth picture description (we use GPT-4 for this), which
we then discard. Then we chop the grid up using PIL, and we're home free.